CR suggestions: -improve style -correct indentation
