+++ package.json
@@ -22,6 +22,7 @@
   "devDependencies": {
     "babel-jest": "23.2.0",
     "babel-preset-react-native": "4.0.0",
+    "eslint": "^5.0.1",
     "jest": "23.2.0",
     "react-test-renderer": "16.3.1"
   },
