+++ src/constants/styleConstants.js
@@ -1,0 +1,2 @@
+
+export const blackColor = 'black';
