+++ src/containers/CreateAccountScreen/styles.js
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1
+  }
+});
+
+export default styles;
