+++ src/containers/LoginScreen/styles.js
@@ -1,0 +1,15 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  redText: {
+    color: 'red',
+    fontSize: 20,
+  }
+});
+
+export default styles;
